services/predictor/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
google-cloud-bigquery==3.25.0
google-cloud-secret-manager==2.20.0
pydantic==2.9.2
db-dtypes==1.3.0
scikit-learn==1.5.2
numpy==2.1.2
joblib==1.4.2